students/16th/team1/kimjunhyung/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
bcrypt==3.2.0 
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
Django==3.1.4
django-cors-headers==3.6.0
django-dotenv==1.4.2
mypy-extensions==0.4.3
mysqlclient==2.0.2
pathspec==0.8.1
pycparser==2.20
PyJWT==2.0.0
pytz==2020.4
regex @ file:///opt/concourse/worker/volumes/live/c84a6349-3315-46e4-634b-b5582dea058b/volume/regex_1606691109605/work
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
